Move to Python3 Support.
It needs PyUSB support (Duh!) but it's been **three** years since people were told that 
Python2 was Sunsetted and DEAD.  While it was a traumatic move for some, it's time
to MOVE ON.  Such as it is with this.  We need the cam to be up and working with
the official driver support because it's at the heart of a PSVR setup and as such, it's
STILL the bargain, full MonadoXR supported VR HMD there is right now.  The only other
REAL full one is the Valve Index at $1k prices.  We have the driver support, but rather 
needed a firmware loader.  This is a udev ruleset capable solution once you add 
PyUSB onto the target system.  Will eventually morph all of this into a proper packaging
solution repository for most systems over time.  It starts with this one fix, though.